moved save/load ohs structures to ohs_settings
