+++ applications/ohs/ohs_settings.cpp
@@ -1,0 +1,17 @@
+#include "ohs_settings.h"
+#include <furi.h>
+#include <file-worker.h>
+
+#define TAG "OHSSettings"
+#define BT_SETTINGS_PATH "/int/ohs.settings"
+
+bool ohs_key_load(Ohs_key* ohs_key){
+    ohs_key.key[28] = {0xf4, 0xa0, 0x49, 0xb5, 0x16, 0xe1, 0xa6, 0xfd, 0x1e, 0x4a,
+                       0x86, 0x62, 0x98, 0x5b, 0xb4, 0x9a, 0x1f, 0x6d, 0xbe, 0x4e,
+                       0xf4, 0xf3, 0x6a, 0xb,  0x39, 0x9b, 0xd9, 0x91};
+    return true;
+}
+
+bool ohs_key_save(Ohs_key* ohs_key){
+    return false;
+}
