+++ applications/ohs/ohs_settings.h
@@ -1,0 +1,8 @@
+#pragma once
+
+typedef struct{
+    uint8_t key[28];
+} Ohs_key;
+
+bool ohs_key_load(Ohs_key* ohs_key);
+bool ohs_key_save(Ohs_key* ohs_key);
